initial commit
This is a sample csv file for pulling in data using cURL.

Note: The .csv file has to be very simple. Use TextWrangler or
something you know can make a plain csv file. Also, make sure it ends
with a hard-return, otherwise it will not read the last line of your
file.